Refactoring InputSources, Pointers, HandDraggable, and misc cleanup. (#2177)

+++ Assets/MixedRealityToolkit/_Core/EventDatum/Input/InputPressedEventData.cs
@@ -17,7 +17,7 @@ namespace Microsoft.MixedReality.Toolkit.Internal.EventDatum.Input
         /// <summary>
         /// The amount, from 0.0 to 1.0, that the select was pressed.
         /// </summary>
-        public double PressedAmount { get; private set; }
+        public float PressedAmount { get; private set; }
 
         /// <inheritdoc />
         public InputPressedEventData(EventSystem eventSystem) : base(eventSystem) { }
@@ -27,7 +27,7 @@ namespace Microsoft.MixedReality.Toolkit.Internal.EventDatum.Input
         /// </summary>
         /// <param name="inputSource"></param>
         /// <param name="pressedAmount"></param>
-        public void Initialize(IMixedRealityInputSource inputSource, double pressedAmount)
+        public void Initialize(IMixedRealityInputSource inputSource, float pressedAmount)
         {
             Initialize(inputSource);
             PressedAmount = pressedAmount;
@@ -39,7 +39,7 @@ namespace Microsoft.MixedReality.Toolkit.Internal.EventDatum.Input
         /// <param name="inputSource"></param>
         /// <param name="handedness"></param>
         /// <param name="pressedAmount"></param>
-        public void Initialize(IMixedRealityInputSource inputSource, Handedness handedness, double pressedAmount)
+        public void Initialize(IMixedRealityInputSource inputSource, Handedness handedness, float pressedAmount)
         {
             Initialize(inputSource, handedness);
             PressedAmount = pressedAmount;
@@ -51,7 +51,7 @@ namespace Microsoft.MixedReality.Toolkit.Internal.EventDatum.Input
         /// <param name="inputSource"></param>
         /// <param name="keyCode"></param>
         /// <param name="pressedAmount"></param>
-        public void Initialize(IMixedRealityInputSource inputSource, KeyCode keyCode, double pressedAmount)
+        public void Initialize(IMixedRealityInputSource inputSource, KeyCode keyCode, float pressedAmount)
         {
             Initialize(inputSource, keyCode);
             PressedAmount = pressedAmount;
@@ -61,9 +61,9 @@ namespace Microsoft.MixedReality.Toolkit.Internal.EventDatum.Input
         /// Populates the event with data.
         /// </summary>
         /// <param name="inputSource"></param>
-        /// <param name="inputType"></param>
+        /// <param name="inputAction"></param>
         /// <param name="pressedAmount"></param>
-        public void Initialize(IMixedRealityInputSource inputSource, InputAction inputAction, double pressedAmount)
+        public void Initialize(IMixedRealityInputSource inputSource, InputAction inputAction, float pressedAmount)
         {
             Initialize(inputSource, inputAction);
             PressedAmount = pressedAmount;
@@ -76,7 +76,7 @@ namespace Microsoft.MixedReality.Toolkit.Internal.EventDatum.Input
         /// <param name="handedness"></param>
         /// <param name="keyCode"></param>
         /// <param name="pressedAmount"></param>
-        public void Initialize(IMixedRealityInputSource inputSource, Handedness handedness, KeyCode keyCode, double pressedAmount)
+        public void Initialize(IMixedRealityInputSource inputSource, Handedness handedness, KeyCode keyCode, float pressedAmount)
         {
             Initialize(inputSource, handedness, keyCode);
             PressedAmount = pressedAmount;
@@ -87,9 +87,9 @@ namespace Microsoft.MixedReality.Toolkit.Internal.EventDatum.Input
         /// </summary>
         /// <param name="inputSource"></param>
         /// <param name="handedness"></param>
-        /// <param name="inputType"></param>
+        /// <param name="inputAction"></param>
         /// <param name="pressedAmount"></param>
-        public void Initialize(IMixedRealityInputSource inputSource, Handedness handedness, InputAction inputAction, double pressedAmount)
+        public void Initialize(IMixedRealityInputSource inputSource, Handedness handedness, InputAction inputAction, float pressedAmount)
         {
             Initialize(inputSource, handedness, inputAction);
             PressedAmount = pressedAmount;
